Refactors table rendering for increased flexibility
Moves table rendering logic to a dedicated function that uses a class to handle HTML attributes.
This change centralizes table configuration and enhances reusability.
Removes obsolete code and tests associated with the previous implementation.
