+++ src/functions/tableHelpers.js
@@ -53,42 +53,3 @@ export function getTableRows(tableColumns, texts, data) {
         return tr;
     });
 }
-
-/**
- * Generates a custom table element for a given part using specified text resources and bindings.
- *
- * @param {Object} part - The data object representing the part to be displayed in the table.
- * @param {Object[]} textResources - An array of text resource objects used for localization.
- * @param {Object} textResourceBindingsForPart - An object containing text resource bindings specific to the part.
- * @param {string} [size="h3"] - The size of the table header (default is "h3").
- * @returns {HTMLElement} A custom table element populated with the provided data and configurations.
- */
-export function getPartTableElement(part, textResources, textResourceBindingsForPart, size = "h3") {
-    const tableColumns = [
-        {
-            titleResourceKey: "col-1",
-            emptyFieldTextResourceKey: "emptyFieldText-default",
-            tagName: "custom-field-part-navn"
-        },
-        {
-            titleResourceKey: "col-2",
-            emptyFieldTextResourceKey: "emptyFieldText-default",
-            tagName: "custom-field-telefonnummer"
-        },
-        {
-            titleResourceKey: "col-3",
-            emptyFieldTextResourceKey: "emptyFieldText-default",
-            tagName: "custom-field-data",
-            dataKey: "epost"
-        }
-    ];
-    const htmlAttributes = new CustomElementHtmlAttributes({
-        formData: { data: part },
-        texts: getTextResourcesFromResourceBindings(textResources, textResourceBindingsForPart),
-        size,
-        hideIfEmpty: true,
-        tableColumns
-    });
-    const tableElement = createCustomElement("custom-table-data", htmlAttributes);
-    return tableElement;
-}
